motion: introduce infrastructure for sub-sampled SAD
When performing a motion search for a block taller than 12 pixels (16 or more)
we want to use sub-sampled SAD, meaning we only measure every other row of the
block.  Normally one would do this by halving the number of rows and doubling
the stride, and doubling the output SAD result but we have two caveats:

1 - our sad_x3 and sad_x4 primitives have hard-coded fenc strides (FENC_STRIDE)
and thus we cannot subsample the fenc block by doubling stride.  Instead we
have to actually subsample the block and store it in a buffer with FENC_STRIDE,
which I've done in this commit.  Now fencSad will always point to a buffer with
FENC_STRIDE, and if this->subsample == 1, the buffer is subsampled.

2 - since the SAD costs never leave this function, we do not have to shift up
our sad results by one bit to account for the subsampling.  The subpel refine
step will remeasure with non-subsampled SATD and thus we don't care what the
best SAD cost actually was (only that it was the best).  The only place where we
need to account for subsampling is in the SAD_THRESH macro.  It must shift the
sad cost up by this->subsample bits.

Steps remaining (all to be done within #if SUBSAMPLE_SAD

1 - Replace fenc with fencSad for all sad(), sad_x3(), sad_x4() calls
2 - shift the luma stride by this->subsample for sad(), sad_x3() and sad_x4()
3 - shift up SAD cost by this->subsample for SAD_THREAD checks
4 - make sure all the changes from 1..3 have no effect on SATD calls
5 - test, we should see a small quality loss and an ME perf gain when enabled