Comment for clarification of the term "cost" used in ME.
In general, we can try to use the following terms in a consistent manner.
1. Distortion: This is the actual distortion produced in the image due to any
of the lossy components in video compression (block matching, quant, filtering etc etc).
In prediction, we can define this as the satd/sad between orig and predicted image.
2. Bits: no. of bits used in representing that particular mode.
3. Cost or TotalCost = Distortion + lambda*bits.

Appending cost to either satd/bits is to be generally avoided.