TEncSlice: ensure lambda per QP does not change [CHANGES OUTPUTS]
This fixes non-determinism when GOP parallelism is enabled. I think we should
just use x264's lambda tables and get rid of all of this logic, but it's too
big of a change to make right before I leave for vacation.

My 1.55 hack attempts to get our compression per-QP close to where it was before
the change.