api: make x265_encoder_get_stats() somewhat future proof
By passing in the size of x265_stats as the user application knows about the
encoder can know not to try to set new fields that were added to the end of
x265_stats.  This requires some discipline on our part to only append to the
structure and to always check the size for any new fields we might add.