nal: allow the concatenated substream buffer to be re-used from frame-to-frame
this removes another malloc/free from the processing of every frame. This commit
also moves the NALList constructor to nal.cpp since it is no longer quite so
trivial. Lastly, it adds a check to prevent crashes in case one or more of the
WPP substreams had a malloc failure.