TEncSearch: fix output non-determination bug
When temporal candidates end up as MVP candidates, they can reach beyond the
available region of valid pixels in the reference picture. In that case, we
should not try to measure distortion there, it only introduces non-determinism.
Just take the other MVP without contest. If both MVP were out of range, we
default to mvpIdx = 0.

Note that there was never any risk of a decode error since when ME runs it
clamps all motion vectors to the range of available pixels.  The setting of
this mvmin/mvmax range is assigned just after picking the best of the two MVP
candidates.