analysis: RDO based BIDIR decisions
At RD 0, 1, and 2, this changes 2Nx2N bidir from a SATD decision to an SA8D
decision.

At RD 3 and 4, if the bidir SA8D cost is within 17/16 of the best inter cost,
then it makes an RDO decision between bestInter and Bidir (allowing psy-rd to
influence the decision, which is the whole point)

At RD 5 and 6, 2Nx2N BIDIR is yet another RD choice at the same level as 2Nx2N
inter and rect and amp. (psy) RDO picks the best mode for each block.