frameencoder: do not use bitmaps for framefilter if not WPP
The non-WPP row loop wants to do frame filter work in between each row, with
a m_filterRowDelay lag. If we use the functions which update the bitmap, it
would allow a worker thread to process a filter row before it was ready.  In
short, the non-WPP path was never intended to work in the presence of a thread
pool. This was causing crashes when --no-wpp --pmode||--pme was used.