motion: use Yuv instance to hold fenc PU pixels (preparing for chroma ME)
This required making an init function which accepts the encoder color space. We
use 4:0:0 for lookahead since it does not keep chroma planes.  Note that I
explicitly renamed this Yuv instance fencPUYuv to make sure people understand it
is not a duplicate of the fencYuv kept by the Analysis structure; it will often
be a sub-partition of the CU fenc yuv.