threadpool: wait until each worker is idle before stopping them
We set the worker's sleep bit when the main loop exits, to make sure the bit is
set if the worker sees m_isActive == false and exits its main loop before the
loop in stop() gets to them.  Without this check, ThreadPool::stop() could
awaken a thread which is not asleep, and this wake event is lost on POSIX. I
encountered one deadlock in the stop() method caused by this.