api: add param.lookaheadSlices
This threading feature used to be auto-enabled when the thread pool was a
certain size. But this makes the encoder output variable depending on the number
of CPUs, which is not particularly helpful, especially when lookahead slices are
often not very helpful with performance.  So I'm putting the choice in the hands
of the user.