encoder: do not disable the thread pool if lookahead-slices is enabled
The user is asking for a pool feature, give them a thread-pool.

This fixes a regression test which used --no-wpp --lookahead-slices 2. By
default this had no thread pool and so lookahead-slices 2 was ignored. But if
the --pme spot-check was randomly selected, it would re-enable the pool and
re-enable lookahead-slices and the outputs would change. After this change,
the lookahead slices prevent --no-wpp from disabling the pool itself.